+++ tests/fuzz/csmith_driver.py
@@ -158,15 +158,6 @@ while 1:
       js_args += ['-s', 'MAIN_MODULE=1']
     if random.random() < 0.25:
       js_args += ['-s', 'INLINING_LIMIT=1'] # inline nothing, for more call interaction
-    if random.random() < 0.01:
-      js_args += ['-s', 'EMTERPRETIFY=1']
-      if random.random() < 0.5:
-        if random.random() < 0.5:
-          js_args += ['-s', 'EMTERPRETIFY_BLACKLIST=["_main"]'] # blacklist main and all inlined into it, but interpret the rest, tests mixing
-        else:
-          js_args += ['-s', 'EMTERPRETIFY_WHITELIST=["_main"]'] # the opposite direction
-      if random.random() < 0.5:
-        js_args += ['-s', 'EMTERPRETIFY_ASYNC=1']
     if random.random() < 0.5:
       js_args += ["--memory-init-file", "0", "-s", "MEM_INIT_METHOD=2"]
     if random.random() < 0.5: